+++ webapp/tests/unit/controller/handlebar-context-caching.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-116: context caching optimization
  * @description     Tests for baseContext caching across nested template expansions
  * @file            webapp/tests/unit/controller/handlebar-context-caching.test.js
- * @version         1.6.21
- * @release         2026-02-21
+ * @version         1.6.22
+ * @release         2026-02-22
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/