+++ webapp/tests/unit/utils/common-pagination.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit Tests for Pagination Utilities
  * @description     Tests for cursor-based and offset-based pagination in CommonUtils
  * @file            webapp/tests/unit/utils/common-pagination.test.js
- * @version         1.6.21
- * @release         2026-02-21
+ * @version         1.6.22
+ * @release         2026-02-22
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/